backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
aiohttp==3.9.1
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
websocket-client==1.6.4
cryptography==41.0.7
psycopg2-binary==2.9.9
alembic==1.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
kiteconnect==5.0.1
APScheduler==3.10.4
